Adapted crates to IOpenable.

+++ src/main/java/cz/tefek/botdiril/userdata/items/crate/ItemCrateEpic.java
@@ -17,7 +17,7 @@ public class ItemCrateEpic extends ItemCrate
     }
 
     @Override
-    public void open(CallObj co, int amount)
+    public void open(CallObj co, long amount)
     {
 
     }
