+++ src/main/java/cz/tefek/botdiril/userdata/items/crate/ItemCrateGlitchy.java
@@ -16,7 +16,7 @@ public class ItemCrateGlitchy extends ItemCrate
     }
 
     @Override
-    public void open(CallObj co, int amount)
+    public void open(CallObj co, long amount)
     {
 
     }
