+++ src/main/java/cz/tefek/botdiril/userdata/items/crate/ItemCrateUncommon.java
@@ -18,7 +18,7 @@ public class ItemCrateUncommon extends ItemCrate
     }
 
     @Override
-    public void open(CallObj co, int amount)
+    public void open(CallObj co, long amount)
     {
 
     }
